change package name dependency to di
