+++ app/src/main/java/com/antonioleiva/mvpexample/app/di/MvpExampleAppModule.java
@@ -1,0 +1,20 @@
+package com.antonioleiva.mvpexample.app.di;
+
+import com.antonioleiva.mvpexample.app.Login.LoginInteractor;
+import com.antonioleiva.mvpexample.app.Login.LoginInteractorImpl;
+
+import dagger.Module;
+import dagger.Provides;
+
+/**
+ * Created by venkatesh on 11/9/15.
+ */
+
+@Module
+public class MvpExampleAppModule {
+
+    @Provides
+    LoginInteractor provideLoginInteractor() {
+        return new LoginInteractorImpl();
+    }
+}
